# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Documentation",
        "description": "Social description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/search/distance": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "Searches users by distance from given point",
                "description": "Searches users by distance from given point",
                "operationId": "searchByDistance",
                "parameters": [
                    {
                        "name": "position",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "latitude": 0,
                            "longitude": 0
                        }
                    },
                    {
                        "name": "minDistance",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "type": "feet",
                            "amount": 0
                        }
                    },
                    {
                        "name": "maxDistance",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "type": "feet",
                            "amount": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login Payload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "password": {
                                            "type": "string"
                                        },
                                        "phone": {
                                            "type": "string"
                                        },
                                        "firstname": {
                                            "type": "string"
                                        },
                                        "lastname": {
                                            "type": "string"
                                        },
                                        "ghostMode": {
                                            "type": "boolean"
                                        },
                                        "subscription": {
                                            "type": "string"
                                        },
                                        "location": {
                                            "properties": {
                                                "latitude": {
                                                    "type": "integer"
                                                },
                                                "longitude": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "locationTimestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "pictureProfile": {
                                            "type": "string"
                                        },
                                        "pictureCover": {
                                            "type": "string"
                                        },
                                        "personalInfo": {
                                            "properties": {
                                                "dob": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "gender": {
                                                    "type": "string"
                                                },
                                                "bio": {
                                                    "type": "string"
                                                },
                                                "education": {
                                                    "type": "string"
                                                },
                                                "politics": {
                                                    "type": "string"
                                                },
                                                "religion": {
                                                    "type": "string"
                                                },
                                                "occupation": {
                                                    "type": "string"
                                                },
                                                "sexuality": {
                                                    "type": "string"
                                                },
                                                "relationship": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "socialInfo": {
                                            "properties": {
                                                "facebook": {
                                                    "type": "string"
                                                },
                                                "instagram": {
                                                    "type": "string"
                                                },
                                                "snapchat": {
                                                    "type": "string"
                                                },
                                                "linkedIn": {
                                                    "type": "string"
                                                },
                                                "twitter": {
                                                    "type": "string"
                                                },
                                                "resume": {
                                                    "type": "string"
                                                },
                                                "coverLetter": {
                                                    "type": "string"
                                                },
                                                "sexuality": {
                                                    "type": "string"
                                                },
                                                "relationship": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "website": {
                                                    "type": "string"
                                                },
                                                "contact": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "liked": {
                                            "type": "boolean"
                                        },
                                        "likeCount": {
                                            "type": "integer"
                                        },
                                        "visitCount": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/search/name": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "Searches users by name",
                "description": "Searches users by name",
                "operationId": "searchByName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}